requirements.txt
camoufox[geoip]
human_requests
aiohttp
aiohttp-retry

pytest
pytest-anyio
anyio
pytest-jsonschema-snapshot
pillow